Create strings it (done the hard way)
A large binary file is provided with the flag somewhere inside.
The goal is to not open the file and find the flag. 
Prior to looking at the hint, I found the flag using this python script.
This outputs a handful of strings surrounded by curly braces.
The flag was clearly identifiable by the "leet-speak-esque" nature of picoCTF flags.
After solving and looking at the hint given, I learned about the "strings" command in the Linux terminal.