Create Nice netcat....
There is a nice program that you can talk to by using this command in a shell: $ nc mercury.picoctf.net 43239, but it doesn't speak English...
The program given feeds back several lines of text, each with an integer. 
I assumed these corresponded to characters so I directly converted to chr in python and it turned into the flag.